Conflicts:
	Assignment3-github/src/com/bgu/assignment3/passives/Orders.java

+++ Assignment3-github/src/com/bgu/assignment3/passives/Dish.java
@@ -24,7 +24,7 @@ public class Dish {
 	private Vector<KitchenTool> kitchenTools;
 	
 	@XmlElement(name = "difficultyRating")
-	private double difficulty;
+	private int difficulty;
 	@XmlElement(name = "reward")
 	private double reward;
